chtype: replace publish pass with chtype -publish_icells
