+++ passes/cmds/Makefile.inc
@@ -57,4 +57,3 @@ OBJS += passes/cmds/abstract.o
 OBJS += passes/cmds/test_select.o
 OBJS += passes/cmds/timeest.o
 OBJS += passes/cmds/linecoverage.o
-OBJS += passes/cmds/publish.o
